pyproject.toml
[build-system]
requires = ['flit_core >=3.9,<4']
build-backend = 'flit_core.buildapi'

[project]
name = "mincepy-sci"
dynamic = ["version"]
description = "Plugins to enable common scientific and machine learning type to be saved by mincePy"
readme = "README.rst"
license = { file = 'LICENSE.txt' }
authors = [
    { name = "Martin Uhrin", email = "martin.uhrin.10@ucl.ac.uk" },
]
keywords = [
    "concurrent",
    "database",
    "nosql",
    "object-store",
    "optimistic-locking",
    "orm",
    "schemaless",
]
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
]
requires-python = '>=3.9'
dependencies = [
    "bidict",
    "mincepy>=0.18.0, <0.19",
]

[project.optional-dependencies]
dev = [
    "pip",
    "pre-commit",
    "prospector",
    "pylint",
    "pytest-cov",
    "pytest>4",
    "twine",
    "yapf",
]
full = [
    "ase",
    "e3nn",
    "flax",
    "jax",
    "numpy",
    "pandas",
    "plams",
    "pyilt22",
    "pymatgen",
    "rdkit",
    "torch",
]

[project.entry-points."mincepy.plugins.types"]
mincepy_sci = "mincepy_sci.provides:get_types"

[project.urls]
Homepage = "https://github.com/muhrin/mincepy_sci.git"

[tool.flit.module]
name = 'mincepy_sci'

[tool.flit.sdist]
exclude = [
    '.github/',
    'docs/',
    'examples/',
    'test/',
]

[tool.flynt]
line-length = 100
fail-on-change = true

[tool.isort]
profile = "black"
force_sort_within_sections = true
include_trailing_comma = true
line_length = 100
multi_line_output = 3

[tool.pylint.format]
max-line-length = 100

[tool.black]
line-length = 100

[tool.pylint.messages_control]
disable = [
    # Unfortunately jaxtyping decorator creates a function that seems to mistakenly be identified as
    # not returning anything, so we have to disable the error below for now
    'assignment-from-no-return',
    'duplicate-code',
    'import-outside-toplevel',
    'missing-docstring',
    'locally-disabled',
    'too-few-public-methods',
    'too-many-instance-attributes',
    'use-dict-literal',
    'arguments-renamed',  # We typically rename positional arguments in subclasses, this is OK
]

[tool.pylint.design]
max-args = 10
max-locals = 20
max-parents = 12
